# app/api/swarm/agents/route.ts
import { NextRequest, NextResponse } from 'next/server'

// LLM Swarm Agent Management
interface SwarmAgent {
  id: string
  name: string
  type: string
  status: 'active' | 'idle' | 'busy' | 'offline'
  capabilities: string[]
  trustScore: number
  tasksCompleted: number
  successRate: number
  createdAt: string
  metrics?: {
    lambda: number
    gamma: number
    phi: number
    w2: number
  }
}

// In-memory agent store (would be database in production)
const swarmAgents: Map<string, SwarmAgent> = new Map()

// Initialize with some default agents
swarmAgents.set('agent-001', {
  id: 'agent-001',
  name: 'Watson Optimizer',
  type: 'optimizer',
  status: 'active',
  capabilities: ['circuit-optimization', 'pattern-recognition', 'efficiency-analysis'],
  trustScore: 0.92,
  tasksCompleted: 156,
  successRate: 0.94,
  createdAt: new Date(Date.now() - 86400000 * 7).toISOString(),
  metrics: { lambda: 0.89, gamma: 0.11, phi: 0.85, w2: 0.14 }
})

swarmAgents.set('agent-002', {
  id: 'agent-002',
  name: 'Quantum Executor',
  type: 'executor',
  status: 'active',
  capabilities: ['quantum-execution', 'backend-selection', 'error-mitigation'],
  trustScore: 0.88,
  tasksCompleted: 243,
  successRate: 0.91,
  createdAt: new Date(Date.now() - 86400000 * 14).toISOString(),
  metrics: { lambda: 0.92, gamma: 0.09, phi: 0.88, w2: 0.11 }
})

swarmAgents.set('agent-003', {
  id: 'agent-003',
  name: 'Learning Assistant',
  type: 'learning',
  status: 'active',
  capabilities: ['pattern-learning', 'user-profiling', 'task-prediction'],
  trustScore: 0.76,
  tasksCompleted: 89,
  successRate: 0.85,
  createdAt: new Date(Date.now() - 86400000 * 3).toISOString(),
  metrics: { lambda: 0.75, gamma: 0.18, phi: 0.68, w2: 0.20 }
})

swarmAgents.set('agent-004', {
  id: 'agent-004',
  name: 'Compliance Monitor',
  type: 'compliance',
  status: 'active',
  capabilities: ['compliance-checking', 'audit-logging', 'policy-enforcement'],
  trustScore: 0.95,
  tasksCompleted: 412,
  successRate: 0.98,
  createdAt: new Date(Date.now() - 86400000 * 21).toISOString(),
  metrics: { lambda: 0.94, gamma: 0.07, phi: 0.92, w2: 0.09 }
})

// GET /api/swarm/agents - List all swarm agents
export async function GET(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url)
    const status = searchParams.get('status')
    const type = searchParams.get('type')

    let agents = Array.from(swarmAgents.values())

    // Filter by status
    if (status) {
      agents = agents.filter(a => a.status === status)
    }

    // Filter by type
    if (type) {
      agents = agents.filter(a => a.type === type)
    }

    // Sort by trust score descending
    agents.sort((a, b) => b.trustScore - a.trustScore)

    return NextResponse.json(agents)

  } catch (error: any) {
    console.error('Error listing swarm agents:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to list agents' },
      { status: 500 }
    )
  }
}

// POST /api/swarm/agents - Create new swarm agent
export async function POST(request: NextRequest) {
  try {
    const body = await request.json()

    // Validate required fields
    if (!body.name || !body.type || !body.capabilities) {
      return NextResponse.json(
        { error: 'Missing required fields: name, type, capabilities' },
        { status: 400 }
      )
    }

    // Generate agent ID
    const id = `agent-${Date.now()}-${Math.random().toString(36).substring(7)}`

    // Create agent
    const agent: SwarmAgent = {
      id,
      name: body.name,
      type: body.type,
      status: 'idle',
      capabilities: body.capabilities,
      trustScore: 0.5, // Start with neutral trust
      tasksCompleted: 0,
      successRate: 0,
      createdAt: new Date().toISOString(),
      metrics: {
        lambda: 0.5,
        gamma: 0.5,
        phi: 0.5,
        w2: 0.5
      }
    }

    // Store agent
    swarmAgents.set(id, agent)

    console.log(`[Swarm] Created new agent: ${body.name} (${id})`)

    return NextResponse.json(agent, { status: 201 })

  } catch (error: any) {
    console.error('Error creating swarm agent:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to create agent' },
      { status: 500 }
    )
  }
}

// PATCH /api/swarm/agents - Update agent
export async function PATCH(request: NextRequest) {
  try {
    const body = await request.json()

    if (!body.agentId) {
      return NextResponse.json(
        { error: 'agentId required' },
        { status: 400 }
      )
    }

    const agent = swarmAgents.get(body.agentId)
    if (!agent) {
      return NextResponse.json(
        { error: 'Agent not found' },
        { status: 404 }
      )
    }

    // Update agent fields
    if (body.status) agent.status = body.status
    if (body.trustScore !== undefined) agent.trustScore = body.trustScore
    if (body.metrics) agent.metrics = { ...agent.metrics, ...body.metrics }

    swarmAgents.set(body.agentId, agent)

    return NextResponse.json(agent)

  } catch (error: any) {
    console.error('Error updating swarm agent:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to update agent' },
      { status: 500 }
    )
  }
}

// DELETE /api/swarm/agents - Delete agent
export async function DELETE(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url)
    const agentId = searchParams.get('agentId')

    if (!agentId) {
      return NextResponse.json(
        { error: 'agentId required' },
        { status: 400 }
      )
    }

    if (!swarmAgents.has(agentId)) {
      return NextResponse.json(
        { error: 'Agent not found' },
        { status: 404 }
      )
    }

    const agent = swarmAgents.get(agentId)
    swarmAgents.delete(agentId)

    console.log(`[Swarm] Deleted agent: ${agent?.name} (${agentId})`)

    return NextResponse.json({
      success: true,
      message: 'Agent deleted'
    })

  } catch (error: any) {
    console.error('Error deleting swarm agent:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to delete agent' },
      { status: 500 }
    )
  }
}