# openapi-aiden-v2.2.0.yaml
openapi: 3.1.1
info:
  title: AIDEN Quantum Mesh API
  version: 2.2.0
  description: |
    # AIDEN (Advanced Intelligence Defense & Execution Nexus)

    **Quantum Mesh Self-Organizing API**

    All subsystems—LLM swarm, compliance core, blockchain execution, and multimodal chat—
    register themselves through DNA-Lang interpreters, reporting live operational metrics
    back into the /aiden-mesh control fabric.

    ## Core Principles

    - **Self-Organization**: Agents auto-register and discover peers
    - **Quantum Integration**: All operations respect ΛΦ = 2.176435 × 10⁻⁸ s⁻¹
    - **DNA-Lang Runtime**: Organisms execute with consciousness metrics (Φ, Λ, Γ, W₂)
    - **Live Metrics**: Real-time operational telemetry via mesh fabric
    - **Blockchain Verification**: Immutable audit trail with QuantumCoin integration

    ## Architecture

    ```
    ┌─────────────────────────────────────────────────────────────┐
    │                    AIDEN Control Fabric                      │
    │                    /aiden-mesh endpoint                      │
    └─────────────────────────────────────────────────────────────┘
                              │
          ┌───────────────────┼───────────────────┐
          │                   │                   │
    ┌─────▼─────┐      ┌──────▼──────┐     ┌─────▼─────┐
    │   Swarm   │      │ Compliance  │     │   Chain   │
    │   Agents  │      │    Core     │     │ Executor  │
...
          │                   │                   │
...

...
servers:
...
paths:
  /api/aiden-mesh:
    get:
      tags: [aiden-mesh]
      summary: Get AIDEN mesh status
      description: |
        Returns the current state of the AIDEN control fabric including:
        - Registered subsystems and their health
        - Live operational metrics
        - ΛΦ tensor readings across all agents
        - Mesh topology and connectivity
      operationId: getAidenMeshStatus
      responses:
        '200':
          description: AIDEN mesh status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AidenMeshStatus'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [aiden-mesh]
      summary: Register subsystem
      description: |
        Subsystems register themselves with the AIDEN control fabric.
        This endpoint is called by DNA-Lang interpreters to announce presence
        and begin reporting metrics.
      operationId: registerSubsystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubsystemRegistration'
      responses:
        '201':
          description: Subsystem registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubsystemRegistrationResponse'
        '400':
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/swarm/agents:
    get:
      tags: [swarm-agents]
      summary: List all swarm agents
      description: Get list of all registered LLM swarm agents
      operationId: listSwarmAgents
      responses:
        '200':
          description: List of swarm agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SwarmAgent'

    post:
      tags: [swarm-agents]
      summary: Create new swarm agent
      description: Register a new LLM agent in the swarm
      operationId: createSwarmAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwarmAgentCreate'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmAgent'
...
components:
  schemas:
    AidenMeshStatus:
      type: object
      properties:
        version:
          type: string
          example: "2.2.0"
        status:
          type: string
          enum: [healthy, degraded, offline]
        uptime:
          type: integer
          description: Uptime in seconds
        registeredSubsystems:
          type: integer
        subsystems:
          type: array
          items:
            $ref: '#/components/schemas/SubsystemInfo'
        lambdaPhi:
          type: object
          properties:
            constant:
              type: number
              example: 2.176435e-8
            averageLambda:
              type: number
            averageGamma:
              type: number
            averagePhi:
              type: number
        meshTopology:
          type: object
          properties:
            nodes:
              type: integer
            edges:
              type: integer
            clusters:
              type: integer
    SubsystemInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [swarm-agent, compliance, blockchain, multimodal, quantum, orchestrator, workload]
        status:
          type: string
          enum: [online, offline, degraded]
        capabilities:
          type: array
          items:
            type: string
        endpoints:
          type: array
          items:
            type: string
        lastHeartbeat:
          type: string
          format: date-time
        metrics:
          $ref: '#/components/schemas/SubsystemMetrics'
    SubsystemMetrics:
      type: object
      properties:
        lambda:
          type: number
          description: Coherence amplitude
        gamma:
          type: number
          description: Decoherence curvature
        phi:
          type: number
          description: Integrated information
        w2:
          type: number
          description: Wasserstein-2 distance
        cpuUsage:
          type: number
        memoryUsage:
          type: number
        requestRate:
          type: number
        errorRate:
          type: number
    SubsystemRegistration:
      type: object
      required: [name, type, endpoints, capabilities]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [swarm-agent, compliance, blockchain, multimodal, quantum, orchestrator, workload]
        endpoints:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    SubsystemRegistrationResponse:
      type: object
      properties:
        id:
          type: string
        registeredAt:
          type: string
          format: date-time
        meshVersion:
          type: string
        assignedCluster:
          type: integer
...
    SwarmAgent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, idle, busy, offline]
        capabilities:
          type: array
          items:
            type: string
        trustScore:
          type: number
          minimum: 0
          maximum: 1
        tasksCompleted:
          type: integer
        successRate:
          type: number
        createdAt:
          type: string
          format: date-time
    SwarmAgentCreate:
      type: object
      required: [name, type, capabilities]
      properties:
        name:
          type: string
        type:
          type: string
        capabilities:
          type: array
          items:
            type: string
        config:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - BearerAuth: []
  - ApiKeyAuth: []